merge main -> prod

+++ api/controllers/test.controllers.js
@@ -23,6 +23,7 @@ const create = async (req, res, next) => {
     instructions,
     scheduledForDate,
     scheduledEndDate,
+    clubId,
   } = req.body;
 
   if (
@@ -36,8 +37,12 @@ const create = async (req, res, next) => {
       message: "1 or more parameter(s) missing from req.body",
     });
   }
-
-  const clubId = req.user.userId;
+  if (clubId != req.user.userId) {
+    return res.status(403).json({
+      message: "This is not your club!",
+    });
+  }
+  // const clubId = req.user.userId;
 
   const test = new Test({
     _id: new mongoose.Types.ObjectId(),
@@ -596,56 +601,63 @@ const addStudents = async (req, res, next) => {
       message: "1 or more parameter(s) missing from req.body",
     });
   }
+  const test = await Test.findById(testId);
 
-  let studentsIdArray = [];
-  const appliedOn = Date.now();
-
-  for (let studentEmail of studentsArray) {
-    let student = await Student.findOneAndUpdate(
-      {
-        email: studentEmail,
-      },
-      {
-        $push: {
-          tests: {
-            testId,
-            clubId,
-            appliedOn,
-            status: "Added/Promoted",
-          },
-        },
-      }
-    );
-    if (student) {
-      let object = {
-        studentId: student.id,
-        marks: 0,
-        corrected: false,
-        responses: [],
-      };
-      studentsIdArray = [...studentsIdArray, object];
-    }
-  }
-  await Test.findOneAndUpdate(
-    { _id: testId },
-    { $addToSet: { users: studentsIdArray } }
-  )
-    .then((result) => {
-      return res.status(200).json({
-        message: "Student array added",
-      });
-    })
-    .catch((err) => {
-      errorLogger.info(
-        `System: ${req.ip} | ${req.method} | ${
-          req.originalUrl
-        } >> ${err.toString()}`
-      );
-      return res.status(500).json({
-        message: "Something went wrong",
-        // error: err.toString(),
-      });
+  if (test.clubId != req.user.userId) {
+    return res.status(403).json({
+      message: "This is not your club!",
     });
+  } else {
+    let studentsIdArray = [];
+    const appliedOn = Date.now();
+
+    for (let studentEmail of studentsArray) {
+      let student = await Student.findOneAndUpdate(
+        {
+          email: studentEmail,
+        },
+        {
+          $push: {
+            tests: {
+              testId,
+              clubId,
+              appliedOn,
+              status: "Added/Promoted",
+            },
+          },
+        }
+      );
+      if (student) {
+        let object = {
+          studentId: student.id,
+          marks: 0,
+          corrected: false,
+          responses: [],
+        };
+        studentsIdArray = [...studentsIdArray, object];
+      }
+    }
+    await Test.findOneAndUpdate(
+      { _id: testId },
+      { $addToSet: { users: studentsIdArray } }
+    )
+      .then((result) => {
+        return res.status(200).json({
+          message: "Student array added",
+        });
+      })
+      .catch((err) => {
+        errorLogger.info(
+          `System: ${req.ip} | ${req.method} | ${
+            req.originalUrl
+          } >> ${err.toString()}`
+        );
+        return res.status(500).json({
+          message: "Something went wrong",
+          // error: err.toString(),
+        });
+      });
+  }
 };
 
 // @desc Publish a test
@@ -656,6 +668,13 @@ const publish = async (req, res, next) => {
   if (!testId) {
     return res.status(400).json({
       message: "1 or more parameter(s) missing from req.query",
+    });
+  }
+  const test = await Test.findById(testId);
+
+  if (test.clubId != req.user.userId) {
+    return res.status(403).json({
+      message: "This is not your club!",
     });
   }
   await Test.findOneAndUpdate({ _id: testId }, { published: true })
@@ -705,6 +724,11 @@ const getAllTestOfAClub = async (req, res, next) => {
   await Test.find({ clubId })
     .select("-usersFinished -usersStarted -users")
     .then(async (tests) => {
+      // if (tests[0].clubId != req.user.userId) {
+      //   return res.status(403).json({
+      //     message: "This is not your club!",
+      //   });
+      // }
       res.status(200).json({
         tests,
       });
@@ -736,6 +760,11 @@ const getAllPublishedTestsOfAClub = async (req, res, next) => {
 
   await Test.find({ clubId, published: true })
     .then(async (tests) => {
+      for (let i = 0; i < tests.length; i++) {
+        if (Date.now() >= tests[i].scheduledEndDate) {
+          tests.splice(i, 1);
+        }
+      }
       res.status(200).json({
         tests,
       });
@@ -769,6 +798,11 @@ const updateTest = async (req, res, next) => {
 
   await Test.findById(testId)
     .then(async (test) => {
+      if (test.clubId != req.user.userId) {
+        return res.status(403).json({
+          message: "This is not your club!",
+        });
+      }
       if (test.scheduledForDate <= Date.now()) {
         return res.status(409).json({
           message: "You can't update the test since it has already started",
@@ -824,7 +858,13 @@ const updateTest = async (req, res, next) => {
 // @route DELETE /api/test/delete
 const deleteTest = async (req, res, next) => {
   const { testId } = req.body;
+  const test = await Test.findById(testId);
 
+  if (test.clubId != req.user.userId) {
+    return res.status(403).json({
+      message: "This is not your club!",
+    });
+  }
   await Question.deleteMany({ testId })
     .then(async () => {
       await Domain.deleteMany({ testId })
@@ -907,4 +947,5 @@ module.exports = {
   getAllTestOfAClub,
   getAllPublishedTestsOfAClub,
   updateTest,
+  deleteTest,
 };
